refactor: 테스트 올바른 위치로 이동
